fix(devstate): cast MAX_HEIGHT to size_t for textPosition in renderHUD

+++ client/gsm/states/scenes/Dev/DevState.cpp
@@ -199,7 +199,8 @@ void DevState::renderHUD() {
     std::string hudText = ss.str();
 
     gfx::color_t white = {255, 255, 255, 255};
-    std::pair<size_t, size_t> textPosition = {10, constants::MAX_HEIGHT - 35};
+    std::pair<size_t, size_t> textPosition =
+        {10, static_cast<size_t>(constants::MAX_HEIGHT - 35)};
     window->drawText(hudText, white, textPosition, "assets/fonts/ARIAL.TTF", 24);
     window->setViewCenter(currentCenter.getX(), currentCenter.getY());
 }
